Finsih cb_order.

+++ modules/cb_schema/migrations/2024-07-20-134140_create_order_table/up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS order_table (
     order_id SERIAL PRIMARY KEY,
     customer_id INTEGER REFERENCES customer_table(customer_id),
     order_date TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-    total_amount DECIMAL(10, 2),
+    total_amount FLOAT,
     status VARCHAR(50) DEFAULT 'PENDING'
 );
 INSERT INTO order_table (customer_id, order_date, total_amount, status) VALUES
